Moved scan button to toolbar
