+++ app/src/main/res/layout/fragment_home.xml
@@ -1,40 +1,19 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical">
 
-    <RelativeLayout
-        android:id="@+id/relative_layout"
-        android:layout_width="match_parent"
-        android:layout_height="75dp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
-
-        <androidx.appcompat.widget.Toolbar
-            android:layout_width="match_parent"
-            android:layout_height="75dp">
-
-            <Button
-                android:id="@+id/btn_scan"
-                android:layout_width="100dp"
-                android:layout_height="50dp"
-                android:text="Scan"></Button>
-        </androidx.appcompat.widget.Toolbar>
-
-    </RelativeLayout>
-
     <ListView
         android:id="@+id/qr_code_list"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
+        android:layout_height="0dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/relative_layout">
-
-    </ListView>
+        app:layout_constraintTop_toTopOf="parent"
+        android:layout_weight="1"/>
 
 
     <androidx.fragment.app.FragmentContainerView
@@ -42,9 +21,6 @@
         android:name="com.example.QArmy.SummaryFragment"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
         tools:layout="@layout/fragment_summary" />
 
-</androidx.constraintlayout.widget.ConstraintLayout>
+</LinearLayout>
